ToArrayParser rename + returned value edition
В связи с этим изменены тесты для класса
Реализован метод listToIntArray, потому что нет assert'а для листов
