+++ src/main/java/calculations/model/utils/ListUtil.java
@@ -1,15 +1,18 @@
 package calculations.model.utils;
 
-public class ToArrayParser {
-    public static int[] parseFromString(String input) {
-        int[] res = new int[input.length()];
+import java.util.ArrayList;
+import java.util.List;
+
+public class ListUtil {
+    public static List<Integer> parseDigitsFromString(String input) {
+        ArrayList<Integer> res = new ArrayList<>(input.length());
 
         for (int i = 0; i < input.length(); i++) {
             int digit = Character.getNumericValue(input.charAt(i));
             if (digit > 9 || digit < 0)
                 throw new IllegalArgumentException("Input string must contain digits only!");
 
-            res[i] = digit;
+            res.add(digit);
         }
 
         return res;
